# Clean up sequence