--- a/lib/sequence.js
+++ b/lib/sequence.js
@@ -1,5 +1,4 @@
 'use strict';
-var Promise = require('bluebird');
 var isPromise = require('is-promise');
 
 module.exports = Sequence;
@@ -9,99 +8,76 @@ function Sequence(tests, bail) {
 		throw new TypeError('Class constructor Sequence cannot be invoked without \'new\'');
 	}
 
+	if (!tests) {
+		throw new Error('Sequence items can\'t be undefined');
+	}
+
+	this.results = [];
+	this.passed = true;
+	this.reason = null;
 	this.tests = tests;
-	this.bail = bail;
+	this.bail = bail || false;
 }
 
 Sequence.prototype.run = function () {
-	var bail = this.bail;
-	var tests = this.tests;
-	var len = tests.length;
-
-	var results = [];
-	var passed = true;
-	var reason = null;
-
-	for (var i = 0; i < len; i++) {
-		var test = tests[i];
-		var result = test.run();
-
-		if (isPromise(result)) {
-			return this._runAsync(results, result, tests.slice(i + 1));
-		}
-
-		results.push(result);
-
-		if (!result.passed) {
-			if (bail) {
-				return {
-					passed: false,
-					result: results,
-					reason: result.reason
-				};
-			} else if (passed) {
-				passed = false;
-				reason = result.reason;
-			}
-		}
-	}
-
-	return {
-		passed: passed,
-		reason: reason,
-		result: results
-	};
+	return this._run();
 };
 
-var BAIL_ERROR = new Error();
+// run sequence items starting from specified index
+Sequence.prototype._run = function (fromIndex) {
+	var length = this.tests.length;
 
-Sequence.prototype._runAsync = function (results, firstAsyncResult, remaining) {
-	var bail = this.bail;
-	var passed = true;
-	var reason = null;
+	if (!fromIndex) {
+		fromIndex = 0;
+	}
 
-	function processResult(result) {
-		results.push(result);
+	for (var i = fromIndex; i < length; i++) {
+		// if last item failed and we should bail, return results and stop
+		if (this.bail && !this.passed) {
+			return this._results();
+		}
 
-		if (!result.passed) {
-			if (passed) {
-				passed = false;
-				reason = result.reason;
-			}
+		var result = this.tests[i].run();
 
-			if (bail) {
-				throw BAIL_ERROR;
-			}
+		if (isPromise(result)) {
+			return this._awaitResult(result, i + 1);
+		}
+
+		this._addResult(result);
+	}
+
+	return this._results();
+};
+
+// if result is a Promise, return it to make Sequence thennable
+// add resolved result and continue execution from the next item
+Sequence.prototype._awaitResult = function (result, nextIndex) {
+	var self = this;
+
+	return result.then(function (ret) {
+		self._addResult(ret);
+
+		return self._run(nextIndex);
+	});
+};
+
+Sequence.prototype._addResult = function (result) {
+	this.results.push(result);
+
+	if (result.passed === false) {
+		this.passed = false;
+
+		// only set reason once
+		if (!this.reason) {
+			this.reason = result.reason;
 		}
 	}
+};
 
-	var result = firstAsyncResult.then(function (firstResult) {
-		processResult(firstResult);
-
-		return Promise.each(remaining, function (test) {
-			var result = test.run();
-
-			if (isPromise(result)) {
-				return result.then(processResult);
-			}
-
-			processResult(result);
-		});
-	});
-
-	if (bail) {
-		result = result.catch(function (err) {
-			if (err !== BAIL_ERROR) {
-				throw err;
-			}
-		});
-	}
-
-	return result.then(function () {
-		return {
-			passed: passed,
-			reason: reason,
-			result: results
-		};
-	});
+Sequence.prototype._results = function () {
+	return {
+		passed: this.passed,
+		reason: this.reason,
+		result: this.results
+	};
 };