@jamestalmage 2016-01-30
This is awesome. Much cleaner than my mess. I need to run a few more tests, but it seems really performant too.