@vdemedes 2016-01-30
PR updated, implemented @jamestalmage's suggestions.