@jamestalmage 2016-01-30
Uh oh :cry: 

```js
// many-async.js
import test from 'ava';

for (var i = 0; i < 10000; i++) {
  test.serial('test' + i, () => {
    return new Promise(resolve => setImmediate(resolve))
  })
}
```

This PR:
```
ava --verbose many-async.js  7.15s user 0.53s system 113% cpu 6.752 total
```

Master:
```
ava --verbose many-async.js  3.76s user 0.46s system 123% cpu 3.421 total
```