@jamestalmage 2016-01-30
```js
// alternating-async.js

import test from 'ava';

for (var i = 0; i < 10000; i++) {
  if (i % 2) {
    test.serial('test' + i, () => {
      return new Promise(resolve => setImmediate(resolve))
    })
  } else {
    test.serial('test' + i, () => {})
  }
}
```

This PR:
```
ava --verbose alternating-async.js  4.52s user 0.44s system 119% cpu 4.158 total
```

Master:
```
ava --verbose alternating-async.js  2.94s user 0.42s system 128% cpu 2.608 total
```