@jamestalmage 2016-01-30
I think `Promise.each` in `bluebird` has some magic we lose out on with this implementation.