@vdemedes 2016-01-30
Ok, I figured it out thanks to iron-node. The fix is amazingly simple - remove this:

```js
Bluebird.longStackTraces();
```

and execution time drops to 2.65s (even faster than master).

It's appeared in this PR, because there's more nesting of promises (no more `Promise.each`) than in master branch.