@jamestalmage 2016-01-30
Hmm. Is there a way to create a promise at runtime that does not use longStackTraces, yet keep longStackTraces turned on for user code?