@jamestalmage 2016-01-31
http://bluebirdjs.com/docs/api/promise.longstacktraces.html

> Long stack traces cannot be disabled after being enabled, and cannot be enabled after promises have alread been created

With longStackTraces disabled, I am getting:

This PR:
```
ava --verbose many-async.js  2.14s user 0.36s system 136% cpu 1.834 total
```

Master:
```
ava --verbose many-async.js  2.24s user 0.36s system 134% cpu 1.930 total
```

This PR is faster (though I'd say a tenth of a second over 10,000 tests is fairly negligible). I think the decision on what to do here ultimately comes down to how much we want to keep longStackTrace support.

I am of two minds:

1. Merge this PR, and work towards a documented `--debug` flag that would:

   - Turn on longStackTraces
   - Set stackTraceLimit to Infinity (meaning we could pick a lower number by default).
   - Set plan stack limit to Infinity
   - Reduce our list of `internals` that we have `stack-utils` filter.
   - Somehow enable the node debugger?? (maybe that's only possible with a single file and the profile scirpt? I don't know, and googling "Node debugging forked processes" is not providing answers).
 
   The current debug strategy (`DEBUG=ava`), is a bit annoying because we also spit out `ipc` and `time-require` stuff that is helpful for contributors, but not really for users. I guess we could still use the `debug` module like two different flags: (`DEBUG=ava-internals` vs just `DEBUG=ava`).

2. Stick with the implementation that we have, since it is faster over a wider array of situations, and leave longStackTraces on. This keeps things simple.

I think I would prefer `1`. Simplicity is good, but longStackTraces have a real impact on performance (for both implementations).