@sindresorhus 2016-02-04
I would go with `1`.