@vdemedes 2016-02-04
> The current debug strategy (DEBUG=ava), is a bit annoying because we also spit out ipc and time-require stuff that is helpful for contributors, but not really for users. I guess we could still use the debug module like two different flags: (DEBUG=ava-internals vs just DEBUG=ava).

There is no need for:

```js
require('debug')('ava-internals')
```

We can do:

 ```js
require('debug')('ava:internals')
```
and

```
$ DEBUG=ava:internals
```

to display only internal logs. Same goes for `ipc` and `time-require`:

```js
require('debug')('ava:ipc') // DEBUG=ava:ipc
require('debug')('ava:require') // DEBUG=ava:require
```

Check out docs on using wildcards for `debug` module - https://www.npmjs.com/package/debug#wildcards.