@jamestalmage 2016-02-05
I understand the wildcard usage. I don't think there is anything special about the `-` vs `:` separator. But, `:` is probably a better choice since that is the convention they have established.

I would just prefer `DEBUG=ava*` not include time-require stuff. So maybe just don't include the `ava` prefix for that? Just call it `time-require`? That is a bit against convention, but `time-require` is not using the `debug` module internally, so I think we can get away with it.

:+1: for finer grained debug scopes (`ava:ipc`, etc).