@vdemedes 2016-02-05
> Just call it time-require

Agree